File Checking is added while file move operation from  Hold folder

+++ JsPlc.Ssc.PetrolPricing/JsPlc.Ssc.PetrolPricing.Portal/Controllers/FileController.cs
@@ -248,7 +248,10 @@ namespace JsPlc.Ssc.PetrolPricing.Portal.Controllers
 		{
 			var heldFile = Path.Combine(_uploadHoldPath, _fileUpload.StoredFileName);
 			var destFile = Path.Combine(_uploadPath, _fileUpload.StoredFileName);
-			File.Move(heldFile, destFile);
+            if (File.Exists(heldFile))
+            {
+                File.Move(heldFile, destFile);
+            }
 			RecordUpload();
 		}
